+++ Threeyes/Plugins/ThreeyesPlugin/Core/Module/Data/FilePathModifier.cs
@@ -4,6 +4,9 @@ using System.IO;
 using UnityEngine;
 namespace Threeyes.Data
 {
+	/// <summary>
+	/// Modify input Path
+	/// </summary>
     public class FilePathModifier
     {
         public string ParentDir { get; set; }
@@ -16,32 +19,8 @@ namespace Threeyes.Data
         }
         public virtual string GetAbsPath(string filePath)
         {
-            //如果传入值是相对路径，就转为全局路径
+            //濡傛灉浼犲叆鍊兼槸鐩稿璺緞锛屽氨杞负鍏ㄥ眬璺緞
             return PathTool.GetAbsPath(ParentDir, filePath);
-        }
-
-        /// <summary>
-        /// Copy file from abs path to relate path, also  file to that path
-        /// </summary>
-        /// <param name="fileAbsPath"></param>
-        /// <returns>relate file Path</returns>
-        public virtual string CopyToRelatePath(string fileAbsPath)
-        {
-            try
-            {
-                string fileName = Path.GetFileName(fileAbsPath);
-                string destDirPath = PathTool.GetOrCreateDir(Path.Combine(ParentDir, SubDir));
-                string destFilePath = Path.Combine(destDirPath, fileName);
-
-                File.Copy(fileAbsPath, destFilePath, true);//Copy file to relate dir(Todo:移动到UIField_FileBase中）
-
-                fileAbsPath = Path.Combine(SubDir, fileName);//Convert to relate path
-            }
-            catch (System.Exception e)
-            {
-                Debug.LogError("Copy file error:\r\n" + e);
-            }
-            return fileAbsPath;
         }
     }
 }