+++ Threeyes/Plugins/ThreeyesPlugin/Core/Module/Persistent/Data/Basic/PersistentData_Color.cs
@@ -1,27 +1,34 @@
 using UnityEngine;
+using Threeyes.Data;
 
 namespace Threeyes.Persistent
 {
-    public class PersistentData_Color : PersistentDataBase<Color, ColorEvent>
-    {
-        private void Reset()
-        {
-            defaultValue = Color.white;
-        }
+	/// <summary>
+	///
+	/// ToAdd:
+	/// -Add HDR support for default/persistent value
+	/// </summary>
+	public class PersistentData_Color : PersistentDataBase<Color, ColorEvent, DataOption_Color>
+	{
+
+		private void Reset()
+		{
+			defaultValue = Color.white;
+		}
 
 #if UNITY_EDITOR
 
-        //¡ª¡ªMenuItem¡ª¡ª
-        static string instName = "ColorPD ";
-        [UnityEditor.MenuItem(strMenuItem_Root_Basic + "Color", false, intBasicMenuOrder + 8)]
-        public static void CreateInst()
-        {
-            Editor.EditorTool.CreateGameObjectAsChild<PersistentData_Color>(instName);
-        }
+		//â€”â€”MenuItemâ€”â€”
+		static string instName = "ColorPD ";
+		[UnityEditor.MenuItem(strMenuItem_Root_Basic + "Color", false, intBasicMenuOrder + 8)]
+		public static void CreateInst()
+		{
+			Editor.EditorTool.CreateGameObjectAsChild<PersistentData_Color>(instName);
+		}
 
-        //¡ª¡ªHierarchy GUI¡ª¡ª
-        public override string ShortTypeName { get { return "Color"; } }
+		//â€”â€”Hierarchy GUIâ€”â€”
+		public override string ShortTypeName { get { return "Color"; } }
 
 #endif
-    }
+	}
 }