+++ Threeyes/SDK/Scripts/Component/BuiltIn/Rendering/AC_SOMaterialControllerConfig.cs
@@ -1,0 +1,5 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+[CreateAssetMenu(menuName = AC_EditorDefinition.AssetMenuPrefix_AC_Config_BuiltIn + "MaterialController", fileName = "MaterialControllerConfig")]
+public class AC_SOMaterialControllerConfig : AC_SOConfigBase<AC_MaterialController.ConfigInfo> { }